20231208 1621 ydj

+++ src/main/java/com/team/creer_back/dto/goods/GoodsReviewDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @Builder  // 클래스 레벨에 붙이거나 생성자에 붙여주면 파라미터를 활용하여 빌더 패턴을 자동으로 생성해줌
 public class GoodsReviewDto {
     private Long goodsReviewId;
-    private Integer goodsDetailId;          // FK-물품 번호 찾기
+    private Long goodsDetailId;          // FK-물품 번호 찾기
     private String nickName;            // FK-구매자 닉네임
     private String memberProfileImg;             // FK-구매자 프사
     private LocalDateTime reviewDate;       // 후기 작성 시각
